First steps to remove NodeExtra namedtuple
The issue with using NodeExtra is that I don't know how to link them
with the actual nodes. So when we move the nodes during transformations
or removal, we need to update the affected gotos/labels' parents while
we only have access to the nodes.

So, I just cloned pycparser, and went through with a hammer and banged
out all of the `__slot__` definitions to allow for monkey patching the
parent attributes in for the Node classes. Terribly clunky, but it
works!