Remove NodeExtra namedtuple
This will remove the last trace of NodeExtra in the code. The namedtuple
has been replaced with the very ugly solution of modifying the pycparser
library to not use the `__slots__` variable. I didn't think to get
memory data before hand, but after removing `__slots__` from pycparser
I'm easily finding the program using over 10MB of data. A better
solution might be adding in *just* the variables that we need into
`__slots__`.